requirements.txt.txt
absl-py==2.1.0
acres==0.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
ajsonrpc==1.2.0
altgraph==0.17.4
amqp==5.3.1
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.7.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
blessed==1.20.0
blinker==1.9.0
bottle==0.13.3
catalogue==2.0.10
celery==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clr_loader==0.2.7.post0
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
contexttimer==0.3.3
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
dlib==19.24.6
et_xmlfile==2.0.0
etelemetry==0.3.1
executing==2.1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.115.6
filelock==3.16.1
fitz==0.0.1.dev2
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.55.2
frozenlist==1.6.0
fsspec==2024.10.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
greenlet==3.2.0
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
httplib2==0.22.0
httptools==0.6.4
huggingface-hub==0.26.5
humanfriendly==10.0
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
ipython==8.30.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jinxed==1.3.0
joblib==1.4.2
kazoo==2.5.0
keras==3.7.0
kiwisolver==1.4.7
kombu==5.5.3
libclang==18.1.1
looseversion==1.3.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.3
matplotlib-inline==0.1.7
maven==0.0.6
mdurl==0.1.2
ml-dtypes==0.4.1
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.4.3
mvn==0.1.3
namex==0.0.8
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.1
numpy==2.0.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pefile==2023.2.7
pillow==11.0.0
platformio==6.1.18
pluggy==1.5.0
prompt_toolkit==3.0.48
propcache==0.3.1
protobuf==5.29.1
prov==2.0.1
proxy-tools==0.1.0
pure_eval==0.2.3
puremagic==1.28
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydot==3.0.3
pyelftools==0.32
PyGetWindow==0.0.9
Pygments==2.18.0
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyJWT==2.10.1
pykafka==2.8.0
PyMsgBox==1.0.9
PyMuPDF==1.25.1
pyngrok==7.2.8
pyparsing==3.2.0
PyPDF2==3.0.1
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pythonnet==3.0.5
pytweening==1.2.0
pytz==2024.2
pywebview==5.4
pywhatkit==5.4
pywin32-ctypes==0.2.3
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
redis==6.0.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
simplejson==3.19.3
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
srsly==2.5.0
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tk==0.1.0
tokenizers==0.21.0
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tqdm==4.67.1
traitlets==5.14.3
traits==6.4.3
transformers==4.47.0
twilio==9.5.2
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.34.0
vine==5.1.0
watchdog==6.0.0
watchfiles==1.0.3
wcwidth==0.2.13
websockets==14.1
Werkzeug==3.1.3
wikipedia==1.4.0
wrapt==1.17.0
wsproto==1.2.0
xlrd==2.0.1
yarl==1.20.0
